% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aws_credentials.R
\name{set_aws_credentials}
\alias{set_aws_credentials}
\title{Install aws credentials in your .Renviron file and
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.}
\usage{
set_aws_credentials(
  keyID,
  secretAccessKey,
  region = "us-east-1",
  install = TRUE
)
}
\arguments{
\item{keyID}{A valid Key ID}

\item{secretAccessKey}{A valid Access key}

\item{region}{A valid AWS region (default set )}

\item{install}{If TRUE, save credentials to .Renviron. Otherwise, set per session.}
}
\value{
keyID, secretAccessKey and region, invisibly
}
\description{
Install aws credentials in your .Renviron file and
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.
}
\examples{

\dontrun{
# Initilialize S3 credentials (this only needs to be done once for a given project)
cori.data.s3::set_aws_credentials(keyID = "###", secretAccessKey = "###")

# Open a duckdb connection configured for S3 access using the credentials
# just installed
con <- connect_to_s3("cori.data.bds")
}

}
